vk: move rasterization code out from vk_render.c
Add vk_raster.{c,h} to handle the traditional 3D rasterization rendering
path.
